Implement crud method and some code refactoring

+++ src/main/java/edu/bada/samochodex/security/auth/ApplicationUserDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface ApplicationUserDao extends CrudRepository<ApplicationUser, Integer> {
+public interface ApplicationUserDao extends CrudRepository<ApplicationUser, Long> {
 
     Optional<ApplicationUser> findByUsername(String username);
 }
